Use Toolkit Search
This change allows users to edit the search engine xml files as they see fit, without it being packed into a JSON. Essentially restoring the way it was in the old days for more user control. The only negative side affect is that we didn't do this sooner, so users may have to re-install any custom search engines.